Solution to second problem: WIP more
This still fails on "(:)", but at least it works on the rest
(and is probably a bit faster since the grammar is now
left-factored).
However, Scala parser combinators have inherently only
support for left-biased choice, so there's no way to make
this work without heavy changes (like switching to GLL-parser
combinators).